Made QuestionDetail component flexible to reuse
